backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
langchain==0.0.346
langchain-community==0.0.1
langchain-google-genai==0.0.2
faiss-cpu==1.7.4
pypdf==3.17.0
sentence-transformers==2.2.2
python-dotenv==1.0.0
google-generativeai==0.3.0